+++ packages/orq-rc/RELEASES.md
@@ -8299,3 +8299,13 @@ Based on:
 - [typescript v4.8.0-rc.14] packages/orq-rc
 ### Releases
 - [NPM v4.8.0-rc.14] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.14 - packages/orq-rc
+
+## 2026-04-14 12:41:55
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.4 (2.879.13) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.8.0-rc.15] packages/orq-rc
+### Releases
+- [NPM v4.8.0-rc.15] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.15 - packages/orq-rc